move query validation
